Return id_token from token endpoint for openid scope
* Add optional nonce to authorization request
* Add CreateIdToken type, called from token endpoint processing function
* If the authorization request contained the openid scope, the id_token
  will be created and added to the token response

A test has been added which uses a dummy token, but still need
to implement creation of the real id_token as a JWT. The algorithms used
(and whether it is encrypted) should depend on the client metadata which
is submitted for the client during registration. If none, then the
default from the registration spec should be used (RS256).