Encode ID Token using a JWK
The OpenID spec (10.1) requires that the ID Token JWT contains the key
ID of the signing key, in the case where the signer has multiple keys.

This change uses the new 'encode' function in jose-jwt 0.3 to sign the
token. The function takes a JWK for the key and if this has a 'kid'
field it is added to the JOSE header.