Support for signed and encrypted UserInfo responses
Adds more information on algorithm preferences to the client data and
makes use of these when creating the response. This also lays the
foundations for better support for id_token and request object preferences.

The latest change in jose-jwt (0.3.1) is required to create the encoded
JWT.

Fixes #2 and test OP-C-04 now passes.