Refactor JWT creation
* Use new Jwt type from jose-jwt
* UserInfo and IdToken creation now use the same shared function, which
encodes based on the client’s algorithm preferences.

This keeps all the complexity related to creating nested tokens and so
on in the one place. The other JWT functions (access and refresh token
creation) could also probably make use of this.

This should hopefully allow issues #3 and #4 to be closed, but needs to
be tested properly.