Introduce server configuration data type
Quite a large refactoring. A first attempt at making the server
functionality more pluggable. The configuration had evolved from the
early prototype and was a bit of a mess with lots of things which need
to be configurable defined inline.

This introduces a Config data type which is initialized with the data
and functions which the server should require. There's also now an
in-memory implementation. The Persist version overwrites the backend
functions with its own.

Test data has been moved out into a separate module, along with the
testBroch function.