Initial changes for port to cryptonite
Makes Handler a MonadRandom instance, which currently just delegates to
IO (i.e. the system DRG). 

Tests will run in ghci, but 'cabal test' will fail because of a linker
error due to the fact that clientsession is still using cipher-aes,
which also exports the '_gmtab' symbol in common with cryptonite,
causing a conflict.