Deal with approval of new scopes
A client may have previously been approved by a user, but make a new
request for additional scopes. The user may also deny some of the scopes
requested.

To deal with this, the denied scopes are now stored with the approval.
If the client makes an authorization request for scopes which are a
subset of the previously approved and denied scopes, the previous
approval will be used and the user will not be prompted again. If the
request includes new scopes the approvals page will again be shown.