Support for 'prompt' parameter
WIP. The user is no longer automatically authenticated before
processAuthorizationRequest is called. Instead an action to obtain
the current user is provided to the function (which may return Nothing)
and processAuthorizationRequest can decide whether authentication should
take place or not based on the prompt parameter value.

If no user is logged in then an error will be returned to the client
in the same way as all the other redirect errors are handled. Otherwise
it will throw a RequiresAuthentication.

TODO: Work out how processAuthorizationRequest can tell whether a
re-authentication has already taken place (when prompt=login).
Currently it will get stuck in a loop because the value will still be
there next time it is called.