Support for 'prompt=login'
The authentication functions now work out whether the user was
authenticated as part of the current authorization request -
authenticatedSubject returns an additional boolean flag to provide
the information, which processAuthorizationRequest can use to decide
whether to force a new login.

When redirecting to the login page, the request URL is first tagged with
a random parameter, which is also posted by the login form. When
the login is successfully processed, it will store the tag in the session.

authenticatedSubject can then retrieve the session value and check if
a query parameter with that name exists. If so, then the current request
caused the login.

User authentication is now split into two configurable functions, one to
return the current user, and one to perform the authentication. This
avoids the hard-coded login URL in brochServer when re-authentication
is required.