Use type-safe redirect URIs
Adds the module Broch.URI which is a wrapper round uri-bytestring.

This introduces a URI type which replaces the Text values which
were used before. The constructor is hidden, so whenever we are dealing
with this type, we know it is based on a valid URI for the
client. It is possible to add a fragment, but the value would then fail
to parse after it had been serialized.

Dealing with sector identifiers is not also handled in this module,
to allow the internal URIRef type to be accessed.