Fix session handling
The case where a session cookie existed at the start of
the request but the session was invalidated during the
request was being ignored. It is logically the same as when
the session from the cookie is found to be expired - the
net result is that if there is no session at the end of the
request, the cookie must be removed.

The function to save the session now takes a flag which
only represents whether the session cookie was present in
the initial request. The code which locates the cookie is
thus now separate from the code to decrypt it, since a
corrupt cookie should also be removed.